+++ demisto_sdk/commands/common/content/objects/pack_objects/pack.py
@@ -1,3 +1,5 @@
+import logging
+import subprocess
 from typing import Any, Iterator, Optional, Union
 
 from demisto_sdk.commands.common.constants import (CLASSIFIERS_DIR,
@@ -28,6 +30,7 @@ from wcmatch.pathlib import Path
 class Pack:
     def __init__(self, path: Union[str, Path]):
         self._path = Path(path)
+        self._metadata = PackMetaData(self._path.joinpath('metadata.json'))
 
     def _content_files_list_generator_factory(self, dir_name: str, suffix: str) -> Iterator[Any]:
         """Generic content objcets iterable generator
@@ -158,6 +161,14 @@ class Pack:
         return obj
 
     @property
+    def metadata(self) -> PackMetaData:
+        return self._metadata
+
+    @metadata.setter
+    def metadata(self, metadata: Optional[PackMetaData]):
+        self._metadata = metadata
+
+    @property
     def secrets_ignore(self) -> Optional[SecretIgnore]:
         obj = None
         file = self._path / ".secrets-ignore"
@@ -192,3 +203,27 @@ class Pack:
             obj = AuthorImage(file)
 
         return obj
+
+    def sign_pack(self, logger: logging.Logger, dumped_pack_dir: Path, sign_directory: Path):
+        """ Signs pack folder and creates signature file.
+
+        Args:
+            logger (logging.Logger): System logger already initialized.
+            dumped_pack_dir (Path): Path to the updated pack to sign.
+            sign_directory (Path): Path to the signDirectory executable file.
+
+        """
+        try:
+            full_command = f'{sign_directory} {dumped_pack_dir} keyfile base64'
+
+            signing_process = subprocess.Popen(full_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+            output, err = signing_process.communicate()
+            signing_process.wait()
+
+            if err:
+                logger.error(f'Failed to sign pack for {self.path.name} - {str(err)}')
+                return
+
+            logger.info(f'Signed {self.path.name} pack successfully')
+        except Exception as error:
+            logger.error(f'Error while trying to sign pack {self.path.name}.\n {error}')