+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
+* Added a validation that makes sure outputs in integrations are matching the README file.
 
 # 1.2.19
 * Fixed an issue where merge id_set was not updated to work with the new entity of Packs.
 * Added a validation that the playbook's version matches the version of its sub-playbooks, scripts, and integrations.
-* Added a validation that makes sure outputs in integrations are matching the README file.
 
 # 1.2.18
 * Changed the *skip-id-set-creation* flag to *create-id-set* in the **validate** command. Its default value will be False.